% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_special_chars.R
\name{add_special_chars}
\alias{add_special_chars}
\title{Add special characters to strings}
\usage{
add_special_chars(data, cols = NULL, messiness = 0.1)
}
\arguments{
\item{data}{input dataframe}

\item{cols}{set of columns to apply transformation to. If \code{NULL}
will apply to all columns. Default \code{NULL}.}

\item{messiness}{Percentage of values to change. Must be
between 0 and 1. Default 0.1.}
}
\value{
a dataframe the same size as the input data.
}
\description{
Add special characters to strings
}
\examples{
add_special_chars(mtcars)
}
