% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_separators.R
\name{alter_separators}
\alias{alter_separators}
\title{Alter separators}
\usage{
alter_separators(data, cols = NULL, messiness = 0.1)
}
\arguments{
\item{data}{input dataframe}

\item{cols}{set of columns to apply transformation to. If \code{NULL}
will apply to all columns. Default \code{NULL}.}

\item{messiness}{Percentage of values to change. Must be
between 0 and 1. Default 0.1.}
}
\value{
a dataframe the same size as the input data.
}
\description{
Randomly alter the separators (spaces and underscores) in character strings
through duplication or replacement
}
\examples{
players <- data.frame(
Player = c("Rey McSriff", "Kevin Nogilny", "Mike Sernandez", "Glenallen Mixon"),
Rating = c("B_3", "A_4", "C_2", "C_3"),
State = c("Arizona", "New Mexico", "North Carolina", "Texas"),
Season = c(2001L, 2002L, 2000L, 2002L)
)
alter_separators(players)
}
