% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_separators.R
\name{change_separators}
\alias{change_separators}
\title{Change separators}
\usage{
change_separators(
  data,
  cols = NULL,
  messiness = 0.1,
  sep_in = c("-", "_", "  ", " "),
  sep_out = c("-", "_", "  ", " ")
)
}
\arguments{
\item{data}{input dataframe}

\item{cols}{set of columns to apply transformation to. If \code{NULL}
will apply to all columns. Default \code{NULL}.}

\item{messiness}{Percentage of values to change. Must be
between 0 and 1. Default 0.1.}

\item{sep_in}{A single value, or vector, or list of what is considered
a separator in the input data. Default \code{c("-", "_", "  ", " ")}.}

\item{sep_out}{A single value, or vector, or list of what the separators
may be randomly with. Default \code{c("-", "_", "  ", " ")}.}
}
\value{
a dataframe the same size as the input data.
}
\description{
Randomly change the separators in character strings
through random replacement
}
\examples{
change_separators(mtcars)
}
