% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_columns.R
\name{duplicate_columns}
\alias{duplicate_columns}
\title{Duplicate columns and insert them into the dataframe at random}
\usage{
duplicate_columns(data, messiness = 0.1, random = TRUE, name_sep = "")
}
\arguments{
\item{data}{input dataframe}

\item{messiness}{Probability that each column is duplicated. Must be
between 0 and 1. Default 0.1.}

\item{random}{Whether duplicated column names should be randomly selected
from other column names, or maintain the original. Default \code{TRUE}.}

\item{name_sep}{Separator to use for adding numbers to end of names. Default \code{""}.}
}
\value{
A dataframe with duplicated rows inserted
}
\description{
Duplicate columns and insert them into the dataframe at random
}
\examples{
duplicate_columns(mtcars, messiness = 0.1)
}
\author{
Jordi Rosell
}
