% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_rows.R
\name{duplicate_rows}
\alias{duplicate_rows}
\title{Duplicate rows and insert them into the dataframe in order or at random. May result in numbers being added to the end of row names.}
\usage{
duplicate_rows(data, messiness = 0.1, shuffle = FALSE)
}
\arguments{
\item{data}{input dataframe}

\item{messiness}{Percentage of rows to duplicate. Must be
between 0 and 1. Default 0.1.}

\item{shuffle}{Insert duplicated data underneath original data or insert randomly}
}
\value{
A dataframe with duplicated rows inserted
}
\description{
Duplicate rows and insert them into the dataframe in order or at random. May result in numbers being added to the end of row names.
}
\examples{
duplicate_rows(mtcars, messiness = 0.1)
}
\author{
Philip Leftwich, Barry Rowlingson
}
