% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_strings.R
\name{make_column_names_messy}
\alias{make_column_names_messy}
\title{Make column names messy}
\usage{
make_column_names_messy(df)
}
\arguments{
\item{df}{data.frame to alter column names}
}
\value{
data.frame with messy column names
}
\description{
Adds special characters and randomly
capitalises characters in the column
names of a data frame.
}
\examples{
make_column_names_messy(iris)
}
