% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_missing.R
\name{make_missing}
\alias{make_missing}
\title{Make missing}
\usage{
make_missing(data, cols = NULL, messiness = 0.1, missing = NA)
}
\arguments{
\item{data}{input dataframe}

\item{cols}{set of columns to apply transformation to. If \code{NULL}
will apply to all columns. Default \code{NULL}.}

\item{messiness}{Percentage of values to change. Must be
between 0 and 1. Default 0.1.}

\item{missing}{A single value, vector, or list of what the
missing values will be replaced with. If length is greater
than 1, values will be replaced randomly.
Default \code{NA}.}
}
\value{
a dataframe the same size as the input data.
}
\description{
Randomly make values missing in all data columns, or a
subset of columns
}
