% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_strings.R
\name{make_string_messy}
\alias{make_string_messy}
\title{Make character vector messy}
\usage{
make_string_messy(x)
}
\arguments{
\item{x}{string vector to mess up}
}
\value{
string vector that is messed up
}
\description{
Adds special characters and randomly
capitalises characters in the provided
character vector.
}
\examples{
make_string_messy(c("Hello", "world"))
}
