% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy.R
\name{messy}
\alias{messy}
\title{Messy}
\usage{
messy(data, messiness = 0.1)
}
\arguments{
\item{data}{input dataframe}

\item{umessiness}{Percentage of values to change per function. Must be
between 0 and 1. Default 0.1.}
}
\value{
a dataframe the same size as the input data.
}
\description{
Make a data frame messier.
}
