% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_colnames.R
\name{messy_colnames}
\alias{messy_colnames}
\title{Make column names messy}
\usage{
messy_colnames(data)
}
\arguments{
\item{data}{data.frame to alter column names}
}
\value{
data.frame with messy column names
}
\description{
Adds special characters and randomly
capitalises characters in the column
names of a data frame.
}
\examples{
messy_colnames(mtcars)
}
