% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_string.R
\name{messy_string}
\alias{messy_string}
\title{messy_string}
\usage{
messy_string(string)
}
\arguments{
\item{string}{A string for which you would like to generate capitalization variants}
}
\value{
A vector of strings
}
\description{
Generate a vector of capitalization variants of a string
}
