% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy.R
\name{messy}
\alias{messy}
\title{Messy}
\usage{
messy(
  data,
  messiness = 0.1,
  missing = NA,
  case_type = "word",
  sep_in = c("-", "_", "  ", " "),
  sep_out = c("-", "_", "  ", " ")
)
}
\arguments{
\item{data}{input dataframe}

\item{messiness}{Percentage of values to change per function. Must be
between 0 and 1. Default 0.1.}

\item{missing}{A single value, vector, or list of what the
missing values will be replaced with. If length is greater
than 1, values will be replaced randomly.
Default \code{NA}.}

\item{case_type}{Whether the case should change based on
the \code{"word"} or \code{"letter"}.}

\item{sep_in}{A single value, or vector, or list of what is considered
a separator in the input data. Default \code{c("-", "_", "  ", " ")}.}

\item{sep_out}{A single value, or vector, or list of what the separators
may be randomly with. Default \code{c("-", "_", "  ", " ")}.}
}
\value{
a dataframe the same size as the input data.
}
\description{
Make a data frame messier.
}
\examples{
messy(mtcars)
}
